Use int64_t for HistogramBase::Count comparison
Fix for cl.exe warning
../../components/browser_watcher/watcher_metrics_provider_win.cc(93): error C2220: warning treated as error - no 'object' file generated
../../components/browser_watcher/watcher_metrics_provider_win.cc(93): warning C4018: '>': signed/unsigned mismatch


Bug:

+++ components/browser_watcher/watcher_metrics_provider_win.cc
@@ -85,7 +85,7 @@ void RecordExitCodes(const base::string16& registry_path) {
   if (!regkey.Valid())
     return;
 
-  size_t num = regkey.GetValueCount();
+  int64_t num = regkey.GetValueCount();
   if (num == 0)
     return;
